POC of python 3.5 async + libcurl
I expect to find that I've done lots of things wrong but as far
as I can tell this gives very good throughput (> 4K requests/s).
It is very possible that it will be able to make this faster
even after fixing all the issues and missing features to make
it comparable to Requests.